+++ plugins/app/src/components/header/HeaderToolbar.tsx
@@ -1,12 +1,6 @@
 import {
-  Avatar,
   Button,
   ButtonVariant,
-  Dropdown,
-  DropdownItem,
-  DropdownSeparator,
-  DropdownToggle,
-  KebabToggle,
   Toolbar,
   ToolbarContent,
   ToolbarGroup,
@@ -14,63 +8,29 @@ import {
   ToolbarItem,
 } from '@patternfly/react-core';
 import React, { useContext, useState } from 'react';
+import { BellIcon } from '@patternfly/react-icons/dist/esm/icons/bell-icon';
 import CogIcon from '@patternfly/react-icons/dist/esm/icons/cog-icon';
 import { Link } from 'react-router-dom';
 import { QuestionCircleIcon } from '@patternfly/react-icons/dist/esm/icons/question-circle-icon';
-import md5 from 'md5';
 
 import { AuthContext, IAuthContext } from '../../context/AuthContext';
+import { INotificationsContext, NotificationsContext } from '../../context/NotificationsContext';
+import HeaderToolbarMobileDropdown from './HeaderToolbarMobileDropdown';
+import HeaderToolbarProfileDropdown from './HeaderToolbarProfileDropdown';
 
-const HeaderToolbar: React.FunctionComponent = () => {
+interface IHeaderToolbarProps {
+  isNotificationDrawerExpanded: boolean;
+  setIsNotificationDrawerExpanded: (value: boolean) => void;
+}
+
+const HeaderToolbar: React.FunctionComponent<IHeaderToolbarProps> = ({
+  isNotificationDrawerExpanded,
+  setIsNotificationDrawerExpanded,
+}: IHeaderToolbarProps) => {
   const authContext = useContext<IAuthContext>(AuthContext);
+  const notificationsContext = useContext<INotificationsContext>(NotificationsContext);
   const [isProfileDrowdownOpen, setIsProfileDrowdownOpen] = useState<boolean>(false);
   const [isMobileDrowdownOpen, setIsMobileDrowdownOpen] = useState<boolean>(false);
-
-  const getProfileImageURL = (): string => {
-    return (
-      'https://secure.gravatar.com/avatar/' + md5(authContext.user.email.toLowerCase().trim()) + '?size=64&default=mm'
-    );
-  };
-
-  const mobileDrowdownOpen: React.ReactElement[] = [];
-  if (authContext.user.email) {
-    mobileDrowdownOpen.push(
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore
-      <DropdownItem key="myprofile" component={(props): React.ReactNode => <Link {...props} to="/profile" />}>
-        My profile
-      </DropdownItem>,
-    );
-    mobileDrowdownOpen.push(<DropdownSeparator key="divider1" />);
-    mobileDrowdownOpen.push(
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore
-      // eslint-disable-next-line jsx-a11y/anchor-has-content
-      <DropdownItem key="logout" component={(props): React.ReactElement => <a {...props} href="/api/auth/logout" />}>
-        Logout
-      </DropdownItem>,
-    );
-    mobileDrowdownOpen.push(<DropdownSeparator key="divider2" />);
-  }
-  mobileDrowdownOpen.push(
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    <DropdownItem key="settings" component={(props): React.ReactElement => <Link {...props} to="/settings" />}>
-      <CogIcon /> Settings
-    </DropdownItem>,
-  );
-  mobileDrowdownOpen.push(
-    <DropdownItem
-      key="help"
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore
-      component={(props): React.ReactElement => (
-        <a {...props} href="https://kobs.io" target="_blank" rel="noreferrer">
-          <QuestionCircleIcon /> Help
-        </a>
-      )}
-    ></DropdownItem>,
-  );
 
   return (
     <Toolbar id="header-toolbar" isFullHeight={true} isStatic={true}>
@@ -80,11 +40,17 @@ const HeaderToolbar: React.FunctionComponent = () => {
           alignment={{ default: 'alignRight' }}
           spacer={{ default: 'spacerNone', md: 'spacerMd' }}
         >
-          {/* <ToolbarItem>
-            <Button aria-label="Notifications" variant={ButtonVariant.plain}>
-              <AttentionBellIcon />
-            </Button>
-          </ToolbarItem> */}
+          {notificationsContext.groups.length > 0 && (
+            <ToolbarItem>
+              <Button
+                aria-label="Notifications"
+                variant={ButtonVariant.plain}
+                onClick={(): void => setIsNotificationDrawerExpanded(!isNotificationDrawerExpanded)}
+              >
+                <BellIcon />
+              </Button>
+            </ToolbarItem>
+          )}
           <ToolbarGroup
             variant={ToolbarGroupVariant['icon-button-group']}
             visibility={{ default: 'hidden', lg: 'visible' }}
@@ -114,49 +80,11 @@ const HeaderToolbar: React.FunctionComponent = () => {
         </ToolbarGroup>
         {/* eslint-disable-next-line @typescript-eslint/naming-convention */}
         <ToolbarItem visibility={{ '2xl': 'hidden', default: 'visible', lg: 'hidden', md: 'hidden', xl: 'hidden' }}>
-          <Dropdown
-            isPlain={true}
-            position="right"
-            toggle={<KebabToggle onToggle={(): void => setIsMobileDrowdownOpen(!isMobileDrowdownOpen)} />}
-            isOpen={isMobileDrowdownOpen}
-            dropdownItems={mobileDrowdownOpen}
-          />
+          <HeaderToolbarMobileDropdown isOpen={isMobileDrowdownOpen} setIsOpen={setIsMobileDrowdownOpen} />
         </ToolbarItem>
         {authContext.user.email ? (
           <ToolbarItem visibility={{ default: 'hidden', md: 'visible' }}>
-            <Dropdown
-              isFullHeight={true}
-              position="right"
-              isOpen={isProfileDrowdownOpen}
-              toggle={
-                <DropdownToggle
-                  icon={<Avatar src={getProfileImageURL()} alt="Avatar" />}
-                  onToggle={(): void => setIsProfileDrowdownOpen(!isProfileDrowdownOpen)}
-                >
-                  {authContext.user.email}
-                </DropdownToggle>
-              }
-              dropdownItems={[
-                <DropdownItem
-                  key="myprofile"
-                  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-                  // @ts-ignore
-                  component={(props): React.ReactElement => <Link {...props} to="/profile" />}
-                >
-                  My profile
-                </DropdownItem>,
-                <DropdownSeparator key="divider" />,
-                <DropdownItem
-                  key="logout"
-                  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-                  // @ts-ignore
-                  // eslint-disable-next-line jsx-a11y/anchor-has-content
-                  component={(props): React.ReactElement => <a {...props} href="/api/auth/logout" />}
-                >
-                  Logout
-                </DropdownItem>,
-              ]}
-            />
+            <HeaderToolbarProfileDropdown isOpen={isProfileDrowdownOpen} setIsOpen={setIsProfileDrowdownOpen} />
           </ToolbarItem>
         ) : null}
       </ToolbarContent>